hide the ASCII_ARMOR symbol
There appears to be some interaction between once_cell and the 32-bit
toolchain for windows that is causing the ASCII_ARMOR symbol to be
mis-linked. To work around the issue reported in #29, hide the
ASCII_ARMOR symbol inside of a function so that it is not even
crate-wide available but is only available inside the function.