allow parsing of PEM files that contain headers
As reported in #36, PEM files can have an (optional) header section after the
BEGIN line and before the base64 encoded data
(https://www.rfc-editor.org/rfc/rfc1421 is a good description of it).
At the very least, we need to support skipping past the headers and retrieving
the base64 encoded data that occurs after them.

To do that, I refactored the parsing code a bit to make it a little more
clear what is going on and make it easy to strip out the headers.
Additionally, I added a little helper to decode the base64 encoded data
so that I can use that in the tests with the expected base64 data so I
can ensure that I am not leaving anything behind when skipping the
headers.